docs: Consolidate API contract unification documentation
- Created unified status & handoff document with all essential information
- Added introduction explaining what API unification is about
- Included quick-start commands for testing and development
- Added completed tasks history and troubleshooting guide
- Consolidated information from 9 separate documents into one
- Removed redundant documentation files
- Added README to guide future developers to the right document
- Moved remove_yasg_safely.py script to docs for reference
- Cleaned up temporary debug scripts from project root
