--- SESSION_HANDOFF_API_CONTRACT.md
@@ -1,42 +1,0 @@
-# Session Handoff – API Contract Unification
-
-## 1 – What We Achieved
-• Backend now **generates OpenAPI 3.1** (`api/openapi.yaml`) every CI run  
-• Frontend **auto-generates a typed client** (`src/api/generated/`) and scripts are in `package.json`  
-• **Cross-repo GitHub Actions** upload the spec & open a *spec-sync* PR in the frontend  
-• Legacy API docs moved to `docs/legacy/` and plan/status docs updated  
-• Documentation added:  
-  • `api_contract_unification_plan.md` – full roadmap & status  
-  • `openapi_issues_to_fix.md` – current schema errors/warnings  
-  • `FACTORY_WORKSPACE_SETUP.md` – draft workspace JSON & instructions  
-
-## 2 – Immediate Next Actions
-1. **Fix Schemathesis auth in backend CI**  
-    Add test credentials or mock auth so contract tests stop returning 401.  
-2. **Resolve frontend type-check failures**  
-    `client/src/pages/inventory.tsx` has syntax errors; correct or exclude from CI.  
-3. **Create `scripts/regenerate_api.sh` & finish Factory workspace**  
-    File watcher will call this script; then create the workspace from the JSON.  
-4. **Eliminate blocking schema errors**  
-    Add `serializer_class` to `CustomObtainAuthToken`, `dev_auth`, `DataEntryViewSet` (see issues doc).
-
-*(All four tasks are needed for green CI and full Section 4/5 completion.)*
-
-## 3 – Key CI Failure Context
-Backend “Validate API contract with Schemathesis” job:  
-  • 392 failures – every call returned **401 Unauthorized** (no credentials).  
-Frontend “type-check” job:  
-  • Eight TypeScript syntax errors in **inventory.tsx** (lines 649-694 & 1692-1694).  
-  • Generated API client compiles fine; failures are unrelated to contract work.
-
-## 4 – Where to Look for Details
-| Topic | File |
-|-------|------|
-| Master roadmap & status | `aquamind/docs/progress/api_contract_unification_plan.md` |
-| Specific OpenAPI errors/warnings | `aquamind/docs/progress/openapi_issues_to_fix.md` |
-| Factory workspace guide & JSON | `FACTORY_WORKSPACE_SETUP.md` |
-
----
-
-You’re set to pick up from here — tackle the four next actions, rerun CI, and the unification project should be ready for merge.  
-Happy hacking! 🚀
