--- aquamind/docs/progress/api-contract-fixes-plan-v2.md
@@ -1,63 +1,0 @@
-# API Contract Fixes – Plan **v2**  
-_AquaMind – Backend_  
-**Created:** 2025-07-21 **Maintainer:** `@Backend-Droid`  
-
----
-
-## 1 · Executive Summary  
-The breakthrough described in `docs/progress/spectacular-migration-findings.md` shows that **drf-spectacular is already the active schema engine**.  
-The only blocker is legacy **`@swagger_auto_schema`** decorators from `drf_yasg`, which suppress spectacular’s auto-detection of query parameters and cause Schemathesis failures.
-
-👉 **Fixing the contract is now far simpler than originally planned (v1).**  
-No manual YAML edits are needed—just clean up decorators, regenerate the spec, and rerun tests.
-
----
-
-## 2 · Task Breakdown  
-
-### A Code Cleanup  
-- [ ] **A1** Search & delete every `@swagger_auto_schema` decorator.  
-- [ ] **A2** Remove related `from drf_yasg...` imports.  
-- [ ] **A3** Commit; ensure app runs (`manage.py check`).
-
-### B Schema Regeneration  
-- [ ] **B1** `python manage.py spectacular --file api/openapi.yaml`  
-- [ ] **B2** `--validate` flag ⇒ confirm zero errors/warnings (aside from optional type-hint notices).  
-- [ ] **B3** Commit regenerated spec.
-
-### C Testing & CI  
-- [ ] **C1** `pytest` – unit & integration suites.  
-- [ ] **C2** `schemathesis run api/openapi.yaml --checks all` – local green run.  
-- [ ] **C3** Push branch ➜ GitHub Actions; contract step must pass.
-
-### D Polish  
-- [ ] **D1** Replace outdated swagger/redoc routes with Spectacular’s `/api/schema/*` in docs & README.  
-- [ ] **D2** Remove `drf-yasg` from `requirements.txt` after CI green.  
-
----
-
-## 3 · Why This Is Easier Than v1  
-| Aspect | v1 Plan | **v2 Simplification** |
-|--------|---------|-----------------------|
-| Query-param docs | Hand-edit YAML for ~50 endpoints | Auto-generated by spectacular once yasg decorators are gone |
-| 400/401/403 codes | Manual additions | Already included by spectacular global settings |
-| Implementation tweaks | Multiple bespoke fixes | Mostly unaffected; only one known 500 bug remains |
-| Total line-changes | Hundreds | ~100 (decorator removals + spec) |
-
-**Estimated effort:** ~1-2 developer days vs. 1+ week in v1.
-
----
-
-## 4 · Success Criteria  
-* OpenAPI 3.1 spec builds with **zero** errors.  
-* Schemathesis contract tests pass in CI.  
-* All existing pytest suites remain green.  
-* `drf-yasg` fully removed; codebase relies solely on **drf-spectacular**.  
-
----
-
-### Revision History  
-
-| Date | Author | Notes |
-|------|--------|-------|
-| 2025-07-21 | Backend-Droid | Drafted v2; supersedes v1 plan |
