--- aquamind/docs/progress/api_contract_unification/CI_TROUBLESHOOTING_LOG.md
@@ -1,72 +1,0 @@
-# CI Troubleshooting Log  
-_AquaMind – API Contract Unification_
-
----
-
-## 1. Overview
-This document records every significant CI failure we hit while unifying the AquaMind API contract and the actions taken to resolve them.  
-Primary focus: the “Validate API contract with Schemathesis” stage inside `AquaMind CI/CD`.
-
-Legend:  
-* ✅ = fixed / no longer occurs  
-* ❌ = still failing / needs work  
-* ⚠️ = mitigated but not root-cause-fixed  
-
----
-
-## 2. Chronological Troubleshooting Attempts
-
-| Date (2025) | Symptom / CI Error | Root Cause | Action / Fix | Result |
-|-------------|-------------------|------------|--------------|--------|
-| **Jul 11** | `django.db.utils.OperationalError` – Windows runner could not apply migrations | Non-ASCII chars in old migration & file-path length limits | Deleted offending migration / squashed, added `.gitattributes` for UTF-8 | ✅ |
-| **Jul 12** | SQLite integer overflow during Schemathesis data generation | OpenAPI `maximum` / `minimum` values exceeded SQLite 64-bit range | Added `clamp_integer_schema_bounds` post-processing hook | ✅ |
-| **Jul 13** | 401 / *ignored_auth* for every endpoint (977 failures) | Spec lacked global `security:` block, so Schemathesis sent no header | Added `ensure_global_security` hook; added `TokenAuthentication` globally | ⚠️ (reduced failures but some persisted) |
-| **Jul 14** | Still ~600 *ignored_auth*; duplicate `tokenAuth` entries clutter spec | DRF-Spectacular duplicated security arrays per op | Added `cleanup_duplicate_security` hook (dedup only) | ⚠️ |
-| **Jul 15** | *ignored_auth* down to 392 but CI user got 403 on admin routes | CI user lacked superuser/staff flags | Management command `get_ci_token` now sets `is_staff` + `is_superuser` | ✅ |
-| **Jul 15** | Pagination tests returning 404 / 500 on `page=0` or huge ints | DRF default paginator raised `NotFound` | Implemented `ValidatedPageNumberPagination` (returns 400, handles OOR) | ✅ |
-| **Jul 16** _(today)_ | *ignored_auth* still ~392; `- {}` anonymous entries in spec | DRF-Spectacular emits `{}` when `AllowAny` present; existed on many ops | Enhanced `cleanup_duplicate_security` to strip `{}` from **all** ops except `/api/v1/auth/token/` & `/dev-auth/`; regenerated spec; committed | ✅ spec cleaned, but CI still fails |
-| **Jul 16** | Schemathesis **status_code_conformance** failures on auth token endpoint (expects 2xx, gets 400) | Fuzzed credentials obviously invalid; schema declares *200 only* | Not solved – need to broaden expected responses or mark as negative test | ❌ |
-| **Jul 16** | 404 on `/api/v1/infrastructure/*` routes | Schemathesis hitting stale paths that were renamed to `/api/v1/batch/*` | URLConf updated previously but **schema** still contains old tags | ❌ awaiting router–schema sync |
-| **Jul 18** | *ignored_auth* ≈ 392 – endpoints returned **200 OK** with invalid auth | **SessionAuthentication bypass** – Schemathesis carried a valid `Cookie:` header, so `SessionAuthentication` authenticated requests before `TokenAuthentication` failed | ① Removed `SessionAuthentication` from `DEFAULT_AUTHENTICATION_CLASSES` (earlier). ② Added explicit `authentication_classes` / `permission_classes` to critical ViewSets (environmental). ③ Helper scripts `add_auth_to_viewsets.py` & `fix_auth_syntax.py` created for bulk updates | ✅ **536 / 537 checks passing (99.8 %)** – only `/api/v1/auth/dev-auth/` false-positive remains |
-| **Jul 18** | 404s on all `/api/v1/infrastructure/*` + 69 unit-test failures | Infrastructure router previously disabled; prune hook dropped infra paths from schema; ViewSets missing auth decorators | a) Re-enabled infrastructure router in `aquamind/api/router.py` (single registration). b) Removed `prune_legacy_paths` from `settings_ci.py` and regenerated schema. c) Added explicit `TokenAuthentication` + `JWTAuthentication` & `IsAuthenticated` to every infrastructure & batch ViewSet. | ✅ **All 69 unit tests PASS**; infra endpoints present & secured, CI green |
-
----
-
-## 3. Current Known CI Issues  — **All Resolved 🎉** (updated 2025-07-21)
-
-
----
-
-
-| Former Issue | Resolution (Jul 21) |
-|--------------|--------------------|
-| Dev-auth endpoint undocumented **401** | Added special-case in `add_standard_responses` hook |
-| Operation-count drift | Schema regenerated after hook cleanup; confirmed **392 operations** expected & present |
-| Log truncation | CI now uploads full `schemathesis-final-test.txt` artefact |
-## 5. Major Breakthrough – 2025-07-18
-
-On July 18 the project hit a decisive milestone:
-
-• **Global Authentication Unified:** `SessionAuthentication` fully removed; every ViewSet now explicitly enforces `TokenAuthentication` + `JWTAuthentication` with `IsAuthenticated` (except deliberate `AllowAny` endpoints).  
-• **Infrastructure Restored & Secured:** Router re-enabled, ViewSets patched, schema regenerated.  
-• **CI Fully Green:** All 69 formerly-failing unit tests now pass, and Schemathesis shows **536 / 537** checks passing (only dev-auth false-positive pending).  
-• **Next Investigation:** Understand why total OpenAPI operations dropped by ~32 % (1716 → 1174) despite endpoint restoration.  
-
-This places the unification effort at **≈99 % completion** – only minor schema whitelisting and operation-count validation remain before final merge.
-
-
-## 4. Lessons Learned
-
-* Always regenerate the OpenAPI schema **after** URL or permission changes; stale paths create false 404s.
-* A top-level `security:` block is mandatory for Schemathesis; rely on a hook to enforce.
-* DRF-Spectacular inserts `{}` whenever any permission class includes `AllowAny`.  
-  – Strip these unless the endpoint is _truly_ public to avoid *ignored_auth*.
-* SQLite contract testing needs integer range clamping; put this in one reusable hook and enable only for CI settings.
-* Give the CI service account superuser rights; otherwise random admin redirects produce unauthorized noise.
-* When Schemathesis fuzzes auth endpoints it will send garbage creds – schema must declare 4xx codes or we must mark operation as negative-test-ignored.
-* **SessionAuthentication is hazardous in APIs** – if a test harness (or browser) accidentally provides a `Cookie:` header, `SessionAuthentication` may override token/JWT checks and mask missing/invalid auth. Removing it globally and enforcing `IsAuthenticated` is the safest posture.
-* Preserve troubleshooting output: redirect `schemathesis run > schemathesis.txt` and upload artefact for easier offline analysis.
-
----
-
-End of log – last updated 2025-07-18
