+++ aquamind/docs/progress/api_contract_unification/README.md
@@ -1,0 +1,33 @@
+# API Contract Unification – README
+
+Welcome!  
+If you’re looking for the latest state of AquaMind’s **API Contract Unification** effort, start with **`UNIFIED_STATUS_AND_HANDOFF.md`** – it is the single source of truth that combines progress, open tasks, troubleshooting and key commands.
+
+---
+
+## Files intentionally kept for reference
+
+| File | Why it’s kept |
+|------|---------------|
+| `UNIFIED_STATUS_AND_HANDOFF.md` | Main status / hand-off document – **read this first** |
+| `remove_yasg_safely.py` | One-off script used to strip legacy *drf-yasg* decorators |
+| `spectacular-migration-findings.md` | Technical notes from initial drf-spectacular migration |
+| `schemathesis_success_summary.md` | Example of a fully green Schemathesis run for future comparison |
+
+---
+
+## Files superseded by the unified document (safe to delete)
+
+* `api_contract_unification_plan.md`
+* `api-contract-fixes-plan.md`
+* `api-contract-fixes-plan-v2.md`
+* `schemathesis_fixes_plan.md`
+* `CI_TROUBLESHOOTING_LOG.md`
+* `session_2025_07_21_ci_schemathesis_handoff.md`
+* `frontend_backend_integration_plan.md`
+* `frontend_backend_integration_checklist.md`
+* `pr_merge_readiness.md`
+
+These older plans and logs have been merged into **`UNIFIED_STATUS_AND_HANDOFF.md`** and are no longer required.
+
+---
