+++ aquamind/docs/progress/api_contract_unification/UNIFIED_STATUS_AND_HANDOFF.md
@@ -1,0 +1,134 @@
+# AquaMind API Contract Unification  
+**Unified Status & Handoff – single-source reference for the next developer / agent**
+
+_Last updated: 2025-07-22_
+
+---
+
+## 📖 Introduction
+AquaMind is moving from multiple ad-hoc API descriptions (Postman, drf-yasg) to **one authoritative OpenAPI 3 spec generated by drf-spectacular**.  
+This “API Contract Unification” project ensures:
+
+* Backend & frontend share a single contract – no more drift.
+* CI validates the contract (Schemathesis) and auto-regenerates the TypeScript client.
+* Legacy swagger-auto-schema decorators (drf-yasg) are removed to avoid dual sources of truth.
+
+This document replaces dozens of troubleshooting logs & plans. It contains _everything_ a new contributor needs: current status, how to run tests, key files, completed history, open tasks and known issues.
+
+---
+
+## 1 Objectives & Current Status
+| Goal | Status | Notes |
+|------|--------|-------|
+| Single **OpenAPI 3** spec via **drf-spectacular** | ✅ `api/openapi.yaml` validated |
+| Remove legacy **drf-yasg** decorators / imports | ✅ 9 files cleaned (91 decorators, 18 imports) |
+| CI job: **OpenAPI generation** | ✅ green |
+| CI job: **Schemathesis contract tests** | ⚠️ **Failing** (401 auth) |
+| Front-end **TS client** regen PR | ⚠️ Partial (workflow stub) |
+| Consolidated docs | 🟢 _this document_ |
+
+---
+
+## 2 Quick-Start Commands 💻
+Run from project root (`AquaMind`):
+
+| Purpose | Command |
+|---------|---------|
+| Activate venv | `.\.venv\Scripts\activate` _(Windows)_ |
+| Django unit tests | `python manage.py test --settings=aquamind.settings_ci` |
+| Run dev server | `python manage.py runserver 0.0.0.0:8000 --settings=aquamind.settings_ci` |
+| Generate / validate spec | `python manage.py spectacular --file api/openapi.yaml --validate` |
+| Get CI token (for Schemathesis) | `python manage.py get_ci_token --settings=aquamind.settings_ci` |
+| **Local Schemathesis** (auth) | `schemathesis run api/openapi.yaml --base-url http://127.0.0.1:8000 --header "Authorization: Token <token>" --checks all` |
+| Run hooks explicitly | `set SCHEMATHESIS_HOOKS=aquamind.utils.schemathesis_hooks` (Win) |
+| Front-end codegen (in frontend repo) | `npm run generate:api` |
+
+---
+
+## 3 Completed Tasks History ✅
+* 2025-07-02 Initial unification plan drafted.
+* 2025-07-10 Enabled drf-spectacular & generated first `openapi.yaml`.
+* 2025-07-15 CI Unicode/SQLite failure resolved.
+* 2025-07-18 Front-end TS client generation script added.
+* 2025-07-21 Regex script `remove_yasg_safely.py` written; all yasg decorators & imports removed without formatting damage.
+* 2025-07-22 Legacy debug scripts deleted, documentation consolidated into this file.
+
+---
+
+## 4 Remaining Tasks 🛠️
+| Priority | Task |
+|----------|------|
+| 🔴 | **Fix Schemathesis CI auth** – generate token, pass as header, strip cookies via hooks. |
+| 🔴 | Update `.github/workflows/contract.yml` to export `AUTH_TOKEN` & set `SCHEMATHESIS_HOOKS`. |
+| 🟠 | Finish front-end spec-sync workflow (`sync-openapi-to-frontend.yml`). |
+| 🟠 | Deprecate & delete legacy Postman / markdown docs in `docs/legacy/`. |
+| 🟢 | Monitor Unicode/SQLite issue (resolved). |
+
+---
+
+## 5 Troubleshooting Guide 🩺
+### Current failing CI symptom
+> Schemathesis: “401 Unauthorized” on most endpoints.
+
+**Root cause**  
+Cassette & live runs included both `Cookie: sessionid` and an invalid `Authorization` header.  
+Locally Django accepts the session cookie → false pass; in CI fresh DB means cookie invalid → 401.
+
+**Fix steps**
+1. Generate fresh token inside workflow:  
+   `python manage.py get_ci_token --settings=aquamind.settings_ci`
+2. Pass _only_ `Authorization: Token <token>` header to Schemathesis.
+3. Ensure `SCHEMATHESIS_HOOKS=aquamind.utils.schemathesis_hooks` – hook strips stray cookies & normalises action responses.
+4. Verify hook logs “AquaMind Schemathesis hooks loaded”.
+
+### Other resolved issues
+| Issue | Resolution |
+|-------|------------|
+| Unicode ✓⚠ℹ in migrations broke SQLite | Replaced with ASCII; set `PYTHONIOENCODING=utf-8` |
+| `get_ci_token` printed nothing | Added `flush=True`, removed premature `sys.exit` |
+
+---
+
+## 6 Important Technical Decisions
+1. **drf-spectacular** = single source of truth.  
+2. **Token / JWT** only – `SessionAuthentication` disabled for REST API.  
+3. Custom Schemathesis hooks (`aquamind.utils.schemathesis_hooks`) manage auth & response quirks.  
+4. Regex decorator removal chosen over AST for stability/formatting.
+
+---
+
+## 7 Key Files & Locations
+| Path | Description |
+|------|-------------|
+| `api/openapi.yaml` | Canonical OpenAPI spec – regenerate via `manage.py spectacular`. |
+| `aquamind/utils/schemathesis_hooks.py` | Hooks: add auth header, strip cookies, fix action responses. |
+| `aquamind/docs/progress/api_contract_unification/remove_yasg_safely.py` | One-off script that removed yasg decorators/imports. |
+| `.github/workflows/contract.yml` | CI job running Schemathesis. _Needs token fix._ |
+| `.github/workflows/sync-openapi-to-frontend.yml` | Auto-opens PR to front-end with regenerated TS client. |
+| `aquamind/settings_ci.py` | CI Django settings incl. `get_ci_token` command. |
+| `Dockerfile.ci` *(if present)* | Container used by GH Actions (Python 3.11, Node 18). |
+
+---
+
+## 8 CI / CD Status 📊
+- **Backend unit tests** – green  
+- **OpenAPI generation** – green  
+- **TypeScript compile** – green (manual)  
+- **Schemathesis** – **red** (auth)  
+- **Spec-sync front-end PR** – workflow exists, not triggered automatically yet
+
+---
+
+## 9 Next-Steps Checklist 🚀
+- [ ] Fix contract workflow auth (see §5).  
+- [ ] Re-run Schemathesis locally; ensure 0 failures.  
+- [ ] Wire front-end sync workflow & confirm PR opens.  
+- [ ] Remove deprecated docs once CI green.  
+- [ ] Merge PR #12.
+
+---
+
+### Contact
+Owner: **@aquarian247**  
+Feel free to ping on GitHub for any clarification.  
+Good luck – you’ve got everything you need here! 🎉
