--- aquamind/docs/progress/api_contract_unification/session_2025_07_15_handoff.md
@@ -1,105 +1,0 @@
-# API-Contract Unification – Daily Handoff  
-*Session*: **2025-07-15**  
-*File*: `session_2025_07_15_handoff.md`  
-*Author*: Code-Droid 🐟  
-
----
-
-## 1  Session Summary & Key Accomplishments
-• Restored green TypeScript build on frontend.  
-• Removed legacy `server/storage.ts` & `routes.ts` in favour of **`server/mock-api.ts`** with env toggle.  
-• Fixed two backend 500s caused by incorrect `search_fields`:  
-  – `MortalityEventViewSet` (`notes` → `description`)  
-  – `JournalEntryViewSet` (`title`, `content` → `description`)  
-• Committed & pushed fixes to `feature/api-contract-unification` in both repos.  
-• Updated `CURRENT_STATUS.md` to reflect new reality.
-
----
-
-## 2  Issues Discovered & Fixed Today
-| # | Component | Symptom | Root Cause | Fix Commit |
-|---|-----------|---------|------------|------------|
-| F-1 | Backend API | Schemathesis crashes on `/batch/mortality-events/` (500) | `search_fields` referenced non-existent `notes` | `f73fea9` |
-| F-2 | Backend API | Schemathesis crashes on `/health/journal-entries/` (500) | `search_fields` referenced `title` & `content` none exist | `f73fea9` |
-| F-3 | Frontend build | Multiple TS errors due to removed `storage.ts` | Legacy imports | removed + mock API drop-in (73be3a3 FE repo) |
-
-All fixes are merged into the feature branch; CI re-runs in progress.
-
----
-
-## 3  Remaining Issues (as of EOD)
-| ID | Area | Status | Notes |
-|----|------|--------|-------|
-| A-1 | Auth enforcement mismatch | **Open** | Schemathesis reports “Authentication declared but not enforced” on several endpoints. Need to audit `permission_classes`. |
-| A-2 | Token capture in CI | **Suspect fixed – needs verify** | Management command now flushes; workflow echoes token length. Watch next run. |
-| A-3 | Pagination edge-cases | **Open** | `page=0` + huge ints cause 404. Decide if spec or impl should change. |
-| X-1 | Docs gap | **Open** | Need Windows Unicode / logging notes, and auth-troubleshooting guide. |
-
----
-
-## 4  Important Discoveries & Patterns
-1. Field-resolution 500s almost always tied to stale `search_fields` / `filterset_fields`. Pattern: docstring updated but config not.  
-2. SQLite CI highlights integer-overflow & auth issues quickly; run `schemathesis` locally with `ci.sqlite3` for parity.  
-3. Token row occasionally created with *empty* key – guard now in `get_ci_token` (recreates token if key falsy).
-
----
-
-## 5  Next Technical Steps
-### Backend
-```text
-1. Verify CI run §“Validate API contract with Schemathesis” prints TOKEN length == 40.
-2. If still 0:
-      • Add `set -e` around token capture
-      • echo token to stderr for debug (mask string).
-3. Auth audit:
-      • grep -R "permission_classes" apps/*/api/viewsets | less
-      • Ensure default `IsAuthenticated` not overridden by allowing anonymous.
-      • Update spec if anon GET intended (add 200/guest security exceptions).
-4. Re-run: `schemathesis run --checks all --base-url=http://127.0.0.1:8000 api/openapi.yaml`.
-5. Pagination:
-      • Confirm `PageNumberPagination` default via `page_query_param`.
-      • If page=0 should be 400 (not 404) update paginator or schema enum minimum=1.
-```
-
-### Frontend
-```text
-• No immediate action. Re-run `npm run generate:api` only after backend spec stabilises.
-```
-
-### Docs
-```text
-• Add Windows logging caveats & auth-enforcement checklist to:
-      docs/quality_assurance/testing_strategy.md
-      docs/quality_assurance/api_security.md
-```
-
----
-
-## 6  Key File Changes Today
-| File | Repo | Purpose |
-|------|------|---------|
-| `apps/batch/api/viewsets.py` | backend | search_fields corrected (`notes` → `description`) |
-| `apps/health/api/viewsets/journal_entry.py` | backend | search_fields corrected (`title`,`content` → `description`) |
-| `apps/users/management/commands/get_ci_token.py` | backend | safeguard against empty token keys + flush |
-| `server/mock-api.ts` | frontend | lightweight mock replacing storage.ts |
-| `server/index.ts` | frontend | conditional routing (mock vs Django) |
-| `aquamind/docs/progress/api_contract_unification/CURRENT_STATUS.md` | backend docs | status table refreshed |
-
----
-
-## 7  CI/CD Status & Error Patterns
-*Last GitHub Actions run* (feature branch, 2025-07-15 18:40 UTC)
-
-Stage | Result | Repeating Errors
------ | ------ | ---------------
-Backend unit tests | ✅ | –
-OpenAPI generation | ✅ | –
-Schemathesis | ❌ | 1) `401 Unauthorized` on many endpoints<br>2) “auth declared but not enforced” warnings<br>3) Remaining pagination 404s
-Frontend build | ✅ | –
-Deploy jobs | ⏳ skipped (blocked on tests)
-
-Expect next run to at least drop 500-error count to zero; focus now on auth consistency.
-
----
-
-*End of hand-off – see you next shift!*
