--- aquamind/docs/progress/api_contract_unification/session_2025_07_16_handoff.md
@@ -1,95 +1,0 @@
-# API-Contract Unification – Daily Handoff  
-*Session*: **2025-07-16**  
-*File*: `session_2025_07_16_handoff.md`  
-*Author*: Code-Droid 🐟  
-
----
-
-## 1  Session Summary & Key Accomplishments
-• Implemented **sustainable** fixes for authentication & pagination – no anonymous endpoints, deterministic page validation.  
-• Added global `tokenAuth` security requirement via `drf-spectacular` + de-duplication hook; regenerated spec.  
-• Introduced `ValidatedPageNumberPagination` (min=1, graceful out-of-range) and made it the DRF default.  
-• CI settings extended with SQLite-safe schema hooks (`clamp_integer_schema_bounds`, `cleanup_duplicate_security`).  
-• Local Schemathesis run passes (392 ops, 0 failures) with auth header.  
-• `CURRENT_STATUS.md` refreshed; CI poised for green run.
-
----
-
-## 2  Issues Discovered & Fixed Today
-
-| # | Component | Symptom | Root Cause | Fix Commit |
-|---|-----------|---------|------------|-----------|
-| F-1 | OpenAPI schema | Schemathesis: “Authentication declared but not enforced” | Spec lacked explicit per-op `security`, duplicates confused tools | `1e45f0b` (global security + deduper) |
-| F-2 | Backend API | 401s on every request in Schemathesis | Token header missing / schema mismatch | Workflow already passed token; spec now enforces single `tokenAuth` |
-| F-3 | Pagination | `page=0` & negative pages gave 404 / 500 | DRF default paginator silently falls through | `3c92b1d` (custom paginator) |
-| F-4 | Schema bloat | Duplicate `security` arrays | drf-spectacular bug | `6b8a8c5` (post-proc hook) |
-
----
-
-## 3  Remaining Issues (EOD)
-
-| ID | Area | Status | Notes |
-|----|------|--------|-------|
-| B-1 | Schemathesis (CI) | **Pending** | Needs fresh GitHub run; expects green. |
-| D-1 | Docs | OpenAPI security & Windows-Unicode caveats | Add to QA guides. |
-| C-1 | Hypothesis flg | `--hypothesis-max-examples=10` still in workflow | Remove once CI stable. |
-
----
-
-## 4  Important Discoveries & Patterns  
-1. Declaring a global `SECURITY` block + dedup hook eliminates “auth not enforced” noise.  
-2. Page validation belongs in code, not spec – returning 400 for invalid pages is clearer than 404.  
-3. SQLite schema clamps + pagination fixes collectively cut Schemathesis runtime by ≈30 %.  
-4. Token row occasionally empty – safeguard added yesterday continues to work (length = 40).
-
----
-
-## 5  Next Technical Steps
-### Backend  
-```text
-1. Push branch to trigger CI; verify Schemathesis ✅.
-2. Remove --hypothesis-max-examples flag to restore full fuzzing depth.
-3. Grep for any remaining custom paginators; consolidate on ValidatedPageNumberPagination.
-4. Document new hooks in api_documentation_standards.md.
-```
-### Frontend  
-```text
-• No action until schema stabilises; then run `npm run generate:api`.
-```
-### Documentation  
-```text
-• Update testing_strategy.md with Windows logging & global-security notes.
-```
-
----
-
-## 6  Key File Changes Today
-
-| File | Purpose |
-|------|---------|
-| `aquamind/utils/pagination.py` | New validated paginator |
-| `aquamind/utils/openapi_utils.py` | Duplicate-security cleanup hook |
-| `aquamind/settings.py` | Global security + paginator default |
-| `aquamind/settings_ci.py` | Added both post-proc hooks |
-| `api/openapi.yaml` | Regenerated with security & pagination metadata |
-| `aquamind/docs/progress/api_contract_unification/CURRENT_STATUS.md` | Status refresh |
-
----
-
-## 7  CI/CD Status & Error Patterns
-*Last GitHub Actions run* (feature branch, **before** today’s fixes)
-
-Stage | Result | Repeating Errors
------ | ------ | ---------------
-Unit tests | ✅ | –
-OpenAPI generation | ✅ | –
-Schemathesis | ❌ | 401 on all ops, “auth declared but not enforced”, page=0 edge
-Frontend build | ✅ | –
-
-*Expect next run* ➜ all stages green. If any failures remain they will likely surface as:  
-• Missing token header (length ≠ 40)  
-• Pagination 400 not reflected in schema (shouldn’t happen after spec regen)
-
----
-
-*End of hand-off – see you next shift!* 🌊
