--- aquamind/docs/progress/api_contract_unification/session_2025_07_17_handoff.md
@@ -1,65 +1,0 @@
-# Session Handoff – **2025-07-17**
-
-## 1. What we did on 2025-07-16
-
-* Investigated persistent Schemathesis “ignored_auth” & status-code failures.
-* Confirmed global `security:` block **is** present in `api/openapi.yaml`.
-* Root cause of lingering anonymous access → operation-level `- {}` entries produced by drf-spectacular.
-* Implemented enhancement in  
-  `aquamind/utils/openapi_utils.py → cleanup_duplicate_security()`  
-  • Removes empty `{}` from *all* operations **except** the two auth endpoints  
-  • Also deduplicates repeated schemes.
-* Regenerated schema & committed:
-  * Only `/api/v1/auth/token/` & `/api/v1/auth/dev-auth/` retain `{}` (intended).
-* Pushed branch `feature/api-contract-unification`.
-
-## 2. Current Schemathesis state
-
-* CI run `16321368520` still **failing** at “Validate API contract with Schemathesis”.
-* Collected ops: 392
-* Earliest failures:
-  * `POST /api/v1/auth/token/` – 400 responses (bad creds fuzz) flagged as **status_code_conformance**.
-  * `GET /api/v1/auth/dev-auth/` – still reported under **ignored_auth** despite correct `{}` in spec.
-  * Several 404s for legacy `/api/v1/infrastructure/…` paths (these endpoints were renamed; routes missing).
-
-## 3. Fixed / verified
-
-| Area | Status |
-|------|--------|
-| Global security block | ✅ present (`[{'tokenAuth': []}]`) |
-| Removal of stray `{}` | ✅ completed via new hook |
-| Pagination edge-case validation | ✅ custom `ValidatedPageNumberPagination` in use |
-| CI user perms & token generation | ✅ user is superuser, token generated (len 40) |
-
-## 4. Still not working
-
-1. Schemathesis still marks many ops as **ignored_auth** even though spec now requires auth. Need to confirm CLI header injection (`Authorization: Token $TOKEN`) works for *every* request.
-2. Status-code failures for auth endpoints (400 ≠ 2xx expected). We may need to allow 400 in schema or configure Schemathesis to skip those.
-3. Numerous 404s for paths under `/api/v1/infrastructure/*` – routes were moved to `/api/v1/batch/*` etc.; schema outdated for router mappings.
-4. CI step exits with code 1 after first failures; no full report presently.
-
-## 5. Next steps
-
-1. **Re-run Schemathesis locally** with `--auth` header and observe which requests miss the header.
-2. Add a **pre-test probe**: log headers received by DRF (`request.META['HTTP_AUTHORIZATION']`) to ensure token arrives.
-3. Update routers / urls:
-   * Remove legacy `infrastructure` URL patterns or update OpenAPI tags.
-4. Consider adding `422`, `400` to acceptable responses for auth endpoints via `@extend_schema`.
-5. If failures persist, capture minimal failing case (`schemathesis reproduce …`) and debug view permissions.
-
-## 6. Key files & locations
-
-* `aquamind/utils/openapi_utils.py` – new hook logic lines ~70-110.
-* `aquamind/settings.py` & `settings_ci.py` – POSTPROCESSING_HOOKS list.
-* `api/openapi.yaml` – regenerated spec (ensure committed).
-* GitHub Action: `.github/workflows/ci.yml` step _Validate API contract with Schemathesis_.
-
-## 7. Blockers / concerns
-
-* **Legacy paths still in schema**: front-end renames not mirrored in URLConf – causes 404 noise.
-* **Schemathesis misclassifying auth**: may require schema tweaks or CLI flags.
-* Timeouts / log truncation make diagnosing CI runs harder; consider uploading `schemathesis-output.txt` artifact.
-
----
-
-_Handed off by previous session.  Please pick up from “Next steps” above. Good luck!_
