--- schemathesis_test_output.txt
@@ -1,65 +1,0 @@
-\U0001f50c AquaMind Schemathesis hooks loaded!
-INFO:schemathesis.hooks:AquaMind Schemathesis hooks initialized
-Warning: Option `--show-errors-tracebacks` is deprecated and will be removed in Schemathesis 4.0. Use `--show-trace` instead
-Warning: Option `--endpoint` is deprecated and will be removed in Schemathesis 4.0. Use `--include-path` instead
-======================= Schemathesis test session starts =======================
-Schema location: file:///C:/Users/bf10087/Projects/AquaMind/api/openapi.yaml
-Base URL: http://127.0.0.1:8000
-Specification version: Open API 3.0.3
-Workers: 1
-Collected API operations: 17
-Collected API links: 0
-API probing: ...
-API probing: SUCCESS
-Schema analysis: ...
-Schema analysis: SKIP
-
-GET /api/v1/auth/dev-auth/ .                                              [  5%]
-GET /api/v1/environmental/readings/recent/ .                              [ 11%]
-POST /api/v1/scenario/scenarios/ .                                        [ 17%]
-GET /api/v1/scenario/scenarios/ .                                         [ 23%]
-POST /api/v1/scenario/scenarios/compare/ .                                [ 29%]
-POST /api/v1/scenario/scenarios/from_batch/ .                             [ 35%]
-GET /api/v1/scenario/scenarios/summary_stats/ .                           [ 41%]
-PUT /api/v1/scenario/scenarios/{scenario_id}/ .                           [ 47%]
-PATCH /api/v1/scenario/scenarios/{scenario_id}/ .                         [ 52%]
-DELETE /api/v1/scenario/scenarios/{scenario_id}/ .                        [ 58%]
-GET /api/v1/scenario/scenarios/{scenario_id}/ .                           [ 64%]
-GET /api/v1/scenario/scenarios/{scenario_id}/chart_data/ .                [ 70%]
-POST /api/v1/scenario/scenarios/{scenario_id}/duplicate/ .                [ 76%]
-GET /api/v1/scenario/scenarios/{scenario_id}/export_projections/ .        [ 82%]
-GET /api/v1/scenario/scenarios/{scenario_id}/projections/ .               [ 88%]
-POST /api/v1/scenario/scenarios/{scenario_id}/run_projection/ .           [ 94%]
-POST /api/v1/scenario/scenarios/{scenario_id}/sensitivity_analysis[...] . [100%]
-
-=================================== SUMMARY ====================================
-
-Performed checks:
-    not_a_server_error                              45 / 45 passed          PASSED 
-    status_code_conformance                         45 / 45 passed          PASSED 
-    content_type_conformance                        45 / 45 passed          PASSED 
-    response_headers_conformance                    45 / 45 passed          PASSED 
-    response_schema_conformance                     45 / 45 passed          PASSED 
-    negative_data_rejection                         45 / 45 passed          PASSED 
-    ignored_auth                                    45 / 45 passed          PASSED 
-
-WARNINGS:
-  - Most of the responses from `GET /api/v1/auth/dev-auth/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `GET /api/v1/environmental/readings/recent/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `POST /api/v1/scenario/scenarios/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `GET /api/v1/scenario/scenarios/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `POST /api/v1/scenario/scenarios/compare/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `POST /api/v1/scenario/scenarios/from_batch/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `GET /api/v1/scenario/scenarios/summary_stats/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `PUT /api/v1/scenario/scenarios/{scenario_id}/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `PATCH /api/v1/scenario/scenarios/{scenario_id}/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `DELETE /api/v1/scenario/scenarios/{scenario_id}/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `GET /api/v1/scenario/scenarios/{scenario_id}/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `GET /api/v1/scenario/scenarios/{scenario_id}/chart_data/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `POST /api/v1/scenario/scenarios/{scenario_id}/duplicate/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `GET /api/v1/scenario/scenarios/{scenario_id}/export_projections/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `GET /api/v1/scenario/scenarios/{scenario_id}/projections/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `POST /api/v1/scenario/scenarios/{scenario_id}/run_projection/` have a 401 status code. Did you specify proper API credentials?
-  - Most of the responses from `POST /api/v1/scenario/scenarios/{scenario_id}/sensitivity_analysis/` have a 401 status code. Did you specify proper API credentials?
-
-============================= 17 passed in 10.23s ==============================
