--- test_api_contract.py
@@ -1,166 +1,0 @@
-import subprocess
-import yaml
-import os
-import sys
-import requests
-
-def run_command(command, cwd=None):
-    """Runs a shell command and returns its output and exit code."""
-    try:
-        process = subprocess.run(
-            command,
-            shell=True,
-            check=False,
-            capture_output=True,
-            text=True,
-            cwd=cwd
-        )
-        return process.stdout, process.stderr, process.returncode
-    except Exception as e:
-        return "", str(e), 1
-
-def test_openapi_generation(backend_repo_path):
-    """
-    Tests if the OpenAPI spec can be generated and reports errors/warnings.
-    """
-    print("--- Testing OpenAPI Spec Generation ---")
-    openapi_file_path = os.path.join(backend_repo_path, "api", "openapi.yaml")
-
-    # Ensure virtual environment is activated for the command
-    command = f"venv\\Scripts\\Activate.ps1; python manage.py spectacular --file api/openapi.yaml"
-    stdout, stderr, exit_code = run_command(command, cwd=backend_repo_path)
-
-    print(f"Command: {command}")
-    print(f"STDOUT:\n{stdout}")
-    print(f"STDERR:\n{stderr}")
-
-    if exit_code == 0:
-        print("[OK] OpenAPI spec generation command executed successfully.")
-    else:
-        print(f"[FAIL] OpenAPI spec generation command failed with exit code {exit_code}.")
-        print("Please check the output above for details.")
-        return False
-
-    # Report errors and warnings from the output
-    errors_found = False
-    warnings_found = False
-    for line in (stdout + stderr).splitlines():
-        if "Error" in line:
-            print(f"[ERROR] Error during generation: {line.strip()}")
-            errors_found = True
-        elif "Warning" in line:
-            print(f"[WARN] Warning during generation: {line.strip()}")
-            warnings_found = True
-
-    if errors_found:
-        print("[FAIL] Errors were found during OpenAPI spec generation. Please review them.")
-        return False
-    elif warnings_found:
-        print("[WARN] Warnings were found during OpenAPI spec generation. Consider addressing them.")
-    else:
-        print("[OK] No errors or warnings reported during OpenAPI spec generation.")
-
-    return True
-
-def validate_yaml_format(backend_repo_path):
-    """
-    Validates if the generated OpenAPI spec is a valid YAML file.
-    """
-    print("\n--- Validating OpenAPI Spec YAML Format ---")
-    openapi_file_path = os.path.join(backend_repo_path, "api", "openapi.yaml")
-
-    if not os.path.exists(openapi_file_path):
-        print(f"❌ OpenAPI spec file not found at: {openapi_file_path}")
-        return False
-
-    try:
-        with open(openapi_file_path, 'r', encoding='utf-8') as f:
-            yaml.safe_load(f)
-        print("[OK] OpenAPI spec is a valid YAML file.")
-        return True
-    except yaml.YAMLError as e:
-        print(f"❌ OpenAPI spec is not a valid YAML file: {e}")
-        return False
-    except Exception as e:
-        print(f"❌ An unexpected error occurred while validating YAML: {e}")
-        return False
-
-def check_key_endpoints_presence(backend_repo_path):
-    """
-    Checks if key endpoints are present in the generated OpenAPI spec.
-    This requires the Django server to be running.
-    """
-    print("\n--- Checking Key Endpoints Presence (Requires Django Server) ---")
-    print("This check is typically performed in CI/CD with Schemathesis.")
-    print("For local testing, ensure your Django development server is running on http://127.0.0.1:8000.")
-
-    openapi_file_path = os.path.join(backend_repo_path, "api", "openapi.yaml")
-    if not os.path.exists(openapi_file_path):
-        print(f"[FAIL] OpenAPI spec file not found at: {openapi_file_path}. Cannot check endpoints.")
-        return False
-
-    try:
-        with open(openapi_file_path, 'r', encoding='utf-8') as f:
-            spec = yaml.safe_load(f)
-    except Exception as e:
-        print(f"❌ Failed to load OpenAPI spec for endpoint check: {e}")
-        return False
-
-    # Define some expected endpoints based on the project structure
-    expected_endpoints = [
-        "/api/v1/users/",
-        "/api/v1/infrastructure/areas/",
-        "/api/v1/batch/batches/",
-        "/api/v1/inventory/feeding-events/",
-        "/api/v1/health/lab-samples/",
-        "/api/schema/", # drf-spectacular schema endpoint
-    ]
-
-    all_endpoints_present = True
-    paths = spec.get('paths', {})
-
-    for endpoint in expected_endpoints:
-        if endpoint in paths:
-            print(f"[OK] Endpoint '{endpoint}' found in OpenAPI spec.")
-        else:
-            print(f"[FAIL] Endpoint '{endpoint}' NOT found in OpenAPI spec.")
-            all_endpoints_present = False
-
-    if all_endpoints_present:
-        print("[OK] All specified key endpoints are present in the OpenAPI spec.")
-    else:
-        print("[FAIL] Some key endpoints are missing from the OpenAPI spec.")
-
-    return all_endpoints_present
-
-def main():
-    backend_repo_path = os.getcwd() # Assuming script is run from backend root
-
-    # Step 1: Test OpenAPI generation and report errors/warnings
-    generation_success = test_openapi_generation(backend_repo_path)
-    if not generation_success:
-        print("\nOverall result: [FAIL] due to OpenAPI generation issues.")
-        sys.exit(1)
-
-    # Step 2: Validate YAML format
-    yaml_valid = validate_yaml_format(backend_repo_path)
-    if not yaml_valid:
-        print("\nOverall result: [FAIL] due to invalid YAML format.")
-        sys.exit(1)
-
-    # Step 3: Check key endpoints presence (requires Django server to be running)
-    # This step is informational for local testing and will be fully validated in CI
-    # with schemathesis.
-    check_key_endpoints_presence(backend_repo_path)
-
-    print("\n--- Local API Contract Test Summary ---")
-    print("OpenAPI Generation: [OK] SUCCESS" if generation_success else "OpenAPI Generation: [FAIL] FAILED")
-    print("YAML Format Validation: [OK] SUCCESS" if yaml_valid else "YAML Format Validation: [FAIL] FAILED")
-    print("Key Endpoints Presence: Check performed (requires running server).")
-    print("\nFor full contract validation, ensure CI/CD pipelines are running Schemathesis.")
-
-    print("\nOverall result: [OK] PASSED (with potential warnings/missing endpoints if server not running).")
-    sys.exit(0)
-
-if __name__ == "__main__":
-    main()
